Sun, Jan 15, 2023, 4:18 AM +00:00

+++ hours/hours.c
@@ -27,8 +27,8 @@ int main(void)
 // TODO: complete the calc_hours function
 float calc_hours(int hours[], int weeks, char output)
 {
-    int sumT = 0;
-    int sumA = 0;
+    float sumT = 0;
+    float sumA = 0;
 
     if (output == 'T')
     {
@@ -44,8 +44,8 @@ float calc_hours(int hours[], int weeks, char output)
         {
             sumA += hours[i];
         }
-        sumA = sumA / (float) weeks;
-        return (float) sumA;
+        sumA =  sumA / weeks;
+        return  sumA;
     }
 
 }
